Updated UI and Cleaned up a bit
Added getCardImagePath() - Which converts the suit and rank into a full image path to print card on UI

Added on to onSubmitBet() - Which allows you to submit bet and get 2 starting cards.

Added onHitButtonClicked() which transfers a random card to the users hand, and checks if the user has busted or not when hit is clicked.

Added displayCardProperly() which allows the card to be properly displayed in the UI when it is moved to the Users hand.

Added displayPlayerHand() which creates the open card placemats for when a card is placed.

Added animateCardToWidget() which is an animation for the cards going towards the widget.