Project CMakeLists Unit Test Toggle
Unit tests can now be toggled on or off in Project Build settings.  If this unit test update is put in your branch, make sure to set BUILD_TESTS to "on" in Project Build settings before commit & pushing.  Build works when tests are off, since tests are not fully completed yet.